getting the data from the tensor

+++ include/ops/init_ops.h
@@ -8,7 +8,7 @@
 void zeros(Tensor* t);
 void ones(Tensor* t);
 void randn(Tensor* t);
-void uniform(Tensor* t, int low, int high);
+void uniform(Tensor* t, float low, float high);
 void from_data(Tensor* t, float* data);
 void borrow(Tensor* t, Storage* data);
